requirements.txt
cohere==1.1.0
flask==2.0.3
langchain-community==0.1.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.1.0
langchain-core==0.1.0
python-dotenv==0.19.2
transformers==4.21.1
sentence-transformers==2.2.0